Move the font related tests from test_basic.py to test_font.py
